# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: CampusFlow API
  version: 2.0.0
  description: >
    Local API for the CampusFlow presentation site. Guests can browse the
    featured profile and public work. Registered users manage only their own
    profile and portfolio. Administrators use a separately configured account.
servers:
  - url: http://localhost:8081
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: campusflow_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
  schemas:
    Account:
      type: object
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string }
        role: { type: string, enum: [USER, ADMIN] }
        provider: { type: string, enum: [local, google, github] }
        enabled: { type: boolean }
    Profile:
      type: object
      properties:
        userId: { type: integer, format: int64 }
        username: { type: string }
        name: { type: string }
        avatar: { type: string }
        title: { type: string }
        summary: { type: string }
        email: { type: string }
        phone: { type: string, description: Returned only to the owner. }
        country: { type: string }
        city: { type: string }
        location: { type: string }
        visibility: { type: string, enum: [PUBLIC, PRIVATE] }
        editable: { type: boolean }
    PortfolioItem:
      type: object
      properties:
        id: { type: integer, format: int64 }
        type: { type: string, enum: [PHOTO, DRAWING, DESIGN, PROJECT, NOTE] }
        title: { type: string }
        description: { type: string }
        imageUrl: { type: string }
        externalUrl: { type: string }
        displayOrder: { type: integer }
        public: { type: boolean }
paths:
  /api/csrf:
    get:
      summary: Start a browser session and obtain the CSRF header value
      responses:
        '200':
          description: CSRF token returned and XSRF-TOKEN cookie set
  /api/register:
    post:
      summary: Register an ordinary user
      description: >
        Public registration always creates a USER account. Passwords must be
        12-72 characters and are stored as BCrypt hashes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, example: demoartist }
                password: { type: string, format: password, minLength: 12, maxLength: 72 }
                displayName: { type: string, example: Kyoto Sketchbook }
      responses:
        '200': { description: Account created and signed in }
        '400': { description: Invalid input }
        '403': { description: Attempted privileged public registration }
        '409': { description: Username already exists }
  /api/authenticate:
    post:
      summary: Sign in with a local account
      description: Sets an HttpOnly campusflow_session cookie after BCrypt verification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        '200': { description: Signed in }
        '401': { description: Unknown, disabled, or incorrect credentials }
  /api/auth/me:
    get:
      summary: Read the current account
      security:
        - sessionCookie: []
      responses:
        '200': { description: Current account or anonymous state }
  /api/auth/logout:
    post:
      summary: End the local session
      security:
        - sessionCookie: []
          csrfHeader: []
      responses:
        '200': { description: Session cookie cleared }
  /api/home:
    get:
      summary: Read Today page data
      description: Uses the supplied location or the featured saved location.
      parameters:
        - name: country
          in: query
          schema: { type: string, example: Japan }
        - name: city
          in: query
          schema: { type: string, example: Kyoto }
      responses:
        '200': { description: Country, city, weather and daily tip data }
  /api/profile:
    get:
      summary: Read the current or featured public profile
      responses:
        '200':
          description: Owner profile when signed in; featured public profile for guests
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
    post:
      summary: Save the signed-in user's own profile and daily location
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Profile' }
      responses:
        '200': { description: Profile saved }
        '401': { description: Sign-in required }
  /api/portfolio:
    get:
      summary: List current-owner or featured public portfolio items
      responses:
        '200': { description: Portfolio items returned }
    post:
      summary: Add an item to the signed-in user's portfolio
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PortfolioItem' }
      responses:
        '200': { description: Item created }
        '401': { description: Sign-in required }
  /api/portfolio/{id}:
    put:
      summary: Update an owned portfolio item
      security:
        - sessionCookie: []
          csrfHeader: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PortfolioItem' }
      responses:
        '200': { description: Item updated }
        '403': { description: Item is owned by another user }
    delete:
      summary: Delete an owned portfolio item
      security:
        - sessionCookie: []
          csrfHeader: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: Item deleted }
        '403': { description: Item is owned by another user }
  /api/admin/users:
    get:
      summary: List accounts for administration
      security:
        - sessionCookie: []
      responses:
        '200': { description: Account list returned }
        '403': { description: Administrator account required }
  /api/admin/users/{id}:
    patch:
      summary: Enable or disable an ordinary account
      security:
        - sessionCookie: []
          csrfHeader: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        '200': { description: Account state updated }
        '403': { description: Administrator account required }